--- a/test/types/timezone.jl
+++ b/test/types/timezone.jl
@@ -1,5 +1,15 @@
 using TimeZones: Class
 
+@testset "TimeZone allocations" begin
+    tz = TimeZone("UTC")  # run once for compilation and to populate cache
+    @assert tz isa FixedTimeZone
+    @test 0 == @allocations(TimeZone("UTC"))
+
+    tz = TimeZone("America/Winnipeg")  # populate cache
+    @assert tz isa VariableTimeZone
+    @test 1 == @allocations(TimeZone("America/Winnipeg"))
+end
+
 @testset "istimezone" begin
     # Invalidate the cache to ensure that `istimezone` works for non-loaded time zones.
     TimeZones._reset_tz_cache()