@nickrobinson251 2022-12-14
```
~/r/TimeZones.jl> TZDATA_VERSION=2016j julia --project=benchmark/ -e 'using PkgBenchmark, TimeZones; export_markdown(stdout, judge(TimeZones, "origin/HEAD", verbose=false))'
PkgBenchmark: Running benchmarks...
[ Info: Installing 2016j tzdata region data
[ Info: Converting tz source files into TimeZone data
PkgBenchmark: creating benchmark tuning file /Users/nickr/repos/TimeZones.jl/benchmark/tune.json...
Tuning 100%|███████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| Time: 0:00:36
Benchmarking 100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| Time: 0:00:43
PkgBenchmark: Running benchmarks...
[ Info: Installing 2016j tzdata region data
[ Info: Converting tz source files into TimeZone data
PkgBenchmark: using benchmark tuning data in /Users/nickr/repos/TimeZones.jl/benchmark/tune.json
Benchmarking 100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| Time: 0:00:48
```
# Benchmark Report for *~/repos/TimeZones.jl*

## Job Properties
* Time of benchmarks:
    - Target: 14 Dec 2022 - 18:05
    - Baseline: 14 Dec 2022 - 18:06
* Package commits:
    - Target: 85eb32
    - Baseline: f8f39b
* Julia commits:
    - Target: 36034a
    - Baseline: 36034a
* Julia command flags:
    - Target: None
    - Baseline: None
* Environment variables:
    - Target: None
    - Baseline: None

## Results
A ratio greater than `1.0` denotes a possible regression (marked with :x:), while a ratio less
than `1.0` denotes a possible improvement (marked with :white_check_mark:). Only significant results - results
that indicate possible regressions or improvements - are shown below (thus, an empty table means that all
benchmark results remained invariant between builds).

| ID                                                           | time ratio                   | memory ratio                 |
|--------------------------------------------------------------|------------------------------|------------------------------|
| `["ZonedDateTime", "local", "standard"]`                     | 0.84 (5%) :white_check_mark: |                   1.00 (1%)  |
| `["ZonedDateTime", "range", "VariableTimeZone/DatePeriod"]`  | 0.91 (5%) :white_check_mark: |                   1.00 (1%)  |
| `["arithmetic", "DatePeriod"]`                               | 0.93 (5%) :white_check_mark: |                   1.00 (1%)  |
| `["arithmetic", "TimePeriod"]`                               |                1.09 (5%) :x: |                   1.00 (1%)  |
| `["arithmetic", "broadcast", "FixedTimeZone/TimePeriod"]`    | 0.94 (5%) :white_check_mark: |                   1.00 (1%)  |
| `["arithmetic", "broadcast", "VariableTimeZone/DatePeriod"]` | 0.93 (5%) :white_check_mark: |                   1.00 (1%)  |
| `["interpret", "local", "ambiguous"]`                        | 0.87 (5%) :white_check_mark: |                   1.00 (1%)  |
| `["interpret", "local", "standard"]`                         | 0.89 (5%) :white_check_mark: |                   1.00 (1%)  |
| `["parse", "ISOZonedDateTimeFormat"]`                        | 0.82 (5%) :white_check_mark: | 0.87 (1%) :white_check_mark: |
| `["parse", "issue#25"]`                                      | 0.92 (5%) :white_check_mark: | 0.88 (1%) :white_check_mark: |
| `["transition_range", "local", "ambiguous"]`                 | 0.83 (5%) :white_check_mark: |                   1.00 (1%)  |
| `["transition_range", "local", "non-existent"]`              | 0.79 (5%) :white_check_mark: |                   1.00 (1%)  |
| `["transition_range", "local", "standard"]`                  | 0.81 (5%) :white_check_mark: |                   1.00 (1%)  |
| `["transition_range", "utc"]`                                |                1.43 (5%) :x: |                   1.00 (1%)  |
| `["tryparsenext_fixedtz", "+06"]`                            | 0.94 (5%) :white_check_mark: |                   1.00 (1%)  |
| `["tryparsenext_fixedtz", "+0600"]`                          | 0.94 (5%) :white_check_mark: |                   1.00 (1%)  |
| `["tryparsenext_fixedtz", "+06:00"]`                         | 0.95 (5%) :white_check_mark: |                   1.00 (1%)  |
| `["tryparsenext_fixedtz", "-06"]`                            | 0.93 (5%) :white_check_mark: |                   1.00 (1%)  |
| `["tryparsenext_fixedtz", "-0600"]`                          | 0.95 (5%) :white_check_mark: |                   1.00 (1%)  |
| `["tryparsenext_fixedtz", "UTC"]`                            | 0.73 (5%) :white_check_mark: |                   1.00 (1%)  |
| `["tryparsenext_fixedtz", "Z"]`                              | 0.56 (5%) :white_check_mark: |                   1.00 (1%)  |
| `["tz_data", "parse_components"]`                            | 0.89 (5%) :white_check_mark: | 0.84 (1%) :white_check_mark: |

## Benchmark Group List
Here's a list of all the benchmark groups executed by this job:

- `["ZonedDateTime"]`
- `["ZonedDateTime", "local"]`
- `["ZonedDateTime", "range"]`
- `["arithmetic"]`
- `["arithmetic", "broadcast"]`
- `["interpret", "local"]`
- `["interpret"]`
- `["parse"]`
- `["transition_range", "local"]`
- `["transition_range"]`
- `["tryparsenext_fixedtz"]`
- `["tryparsenext_tz"]`
- `["tz_data"]`

## Julia versioninfo

### Target
```
Julia Version 1.8.2
Commit 36034abf260 (2022-09-29 15:21 UTC)
Platform Info:
  OS: macOS (x86_64-apple-darwin21.4.0)
  uname: Darwin 21.4.0 Darwin Kernel Version 21.4.0: Mon Feb 21 20:35:58 PST 2022; root:xnu-8020.101.4~2/RELEASE_ARM64_T6000 x86_64 i386
  CPU: Apple M1 Max:
                 speed         user         nice          sys         idle          irq
       #1-10  2400 MHz    1580681 s          0 s    1469385 s   30832641 s          0 s
  Memory: 64.0 GB (2506.71875 MB free)
  Uptime: 446789.0 sec
  Load Avg:  2.73779296875  2.6376953125  2.58837890625
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-13.0.1 (ORCJIT, westmere)
  Threads: 1 on 10 virtual cores
```

### Baseline
```
Julia Version 1.8.2
Commit 36034abf260 (2022-09-29 15:21 UTC)
Platform Info:
  OS: macOS (x86_64-apple-darwin21.4.0)
  uname: Darwin 21.4.0 Darwin Kernel Version 21.4.0: Mon Feb 21 20:35:58 PST 2022; root:xnu-8020.101.4~2/RELEASE_ARM64_T6000 x86_64 i386
  CPU: Apple M1 Max:
                 speed         user         nice          sys         idle          irq
       #1-10  2400 MHz    1581658 s          0 s    1469753 s   30837878 s          0 s
  Memory: 64.0 GB (2527.17578125 MB free)
  Uptime: 446854.0 sec
  Load Avg:  2.205078125  2.5078125  2.54345703125
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-13.0.1 (ORCJIT, westmere)
  Threads: 1 on 10 virtual cores